moved attachment part of submit to attachmentservice
