+++ backend/src/main/java/com/synbiohub/sbh3/dao/SparqlService.java
@@ -4,11 +4,13 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.synbiohub.sbh3.repo.SparqlRepository;
 import com.synbiohub.sbh3.sparql.SPARQLQuery;
+import com.synbiohub.sbh3.utils.StringUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.HashMap;
 import java.util.Map;
 
 @RequiredArgsConstructor
@@ -56,6 +58,42 @@ public class SparqlService {
                 break;
             }
         }
+    }
+
+    /** Maps existing {@code sbol:source} values (e.g. {@code file:foo.png}) to attachment URIs. */
+    public Map<String, String> loadAttachmentSources(String collectionUri, String graphUri) throws IOException {
+        String query = new SPARQLQuery(SparqlRepository.GET_ATTACHMENT_SOURCE_SPARQL)
+                .loadTemplate(Map.of("uri", collectionUri));
+        String raw = sparqlRepository.getQuery(query, graphUri);
+        Map<String, String> sources = new HashMap<>();
+        JsonNode bindings = objectMapper.readTree(raw).path("results").path("bindings");
+        if (!bindings.isArray()) {
+            return sources;
+        }
+        for (JsonNode row : bindings) {
+            String source = row.path("source").path("value").asText(null);
+            String attachment = row.path("attachment").path("value").asText(null);
+            if (source != null && attachment != null) {
+                sources.put(source, attachment);
+            }
+        }
+        return sources;
+    }
+
+    public void attachUpload(Map<String, String> params, String graphUri, boolean jsonResults) throws IOException {
+        String query = new SPARQLQuery(sparqlRepository.ATTACH_UPLOAD_SPARQL).loadTemplate(params);
+        update(query, graphUri, false);
+    }
+
+    /** Replaces hash/size on an existing attachment when re-uploading the same {@code file:} source. */
+    public void updateAttachment(String graphUri, String attachmentUri, String uploadHash, long size)
+            throws IOException {
+        String query = new SPARQLQuery(sparqlRepository.UPDATE_ATTACHMENT_SPARQL).loadTemplate(Map.of(
+                "attachmentURI", attachmentUri,
+                "attachmentSource", attachmentUri + "/download",
+                "hash", StringUtil.sparqlStringLiteral(uploadHash),
+                "size", StringUtil.sparqlStringLiteral(Long.toString(size))));
+        update(query, graphUri, false);
     }
 
 }
